Backend/requirements.txt
flask
pymongo
pyjwt
bcrypt
pycryptodome
opencv-python
